+++ apps/docs/src/server/content/tutorial.md
@@ -8,19 +8,13 @@ The following documentation covers the basics of creating a site and all of the 
 
 To get started, you'll need to have [Node](https://nodejs.org) (recommended), [Bun](https://bun.sh/), or [Deno](https://deno.com) or installed on your computer. Then run the [`create-domco`](https://github.com/rossrobino/domco/tree/main/packages/create-domco) script to create a new project. If you already have an existing client-side Vite project check out the [migration instructions](/migrate).
 
-### Node
-
 ```bash
 npm create domco@__CREATE_VERSION__
 ```
 
-### Bun
-
 ```bash
 bun create domco@__CREATE_VERSION__
 ```
-
-### Deno
 
 ```bash
 deno run -A npm:create-domco@__CREATE_VERSION__
@@ -106,7 +100,7 @@ domco takes care of these problems using [virtual modules](https://vitejs.dev/gu
 
 ### client:page
 
-You can import the transformed HTML of any `+page.html` from this module or one of it's sub-paths. domco calls Vite's [`transformIndexHtml`](https://vitejs.dev/guide/api-plugin.html#transformindexhtml) hook on the imported page and inlines it into your server build.
+You can import the transformed HTML of any `+page.html` from this module or one of it's sub-paths. domco calls Vite's [`transformIndexHtml`](https://vitejs.dev/guide/api-plugin.html#transformindexhtml) hook on the imported page and inlines it into your server bundle.
 
 ```ts {3,10}
 // src/server/+func.ts
@@ -133,7 +127,7 @@ You can also easily get the `<script>` tags for any `+script` module on the serv
 
 In development, domco links the scripts to the source. In production, domco [reads](https://vitejs.dev/guide/backend-integration.html) the [manifest](https://vitejs.dev/config/build-options.html#build-manifest) generated by the client build and includes the hashed version of these file names and their imports.
 
-```ts {3,14}
+```ts {3,15}
 // src/server/+func.ts
 // returns transformed content of `src/client/+script.ts`
 import { tags } from "client:script";
@@ -163,7 +157,7 @@ export default {
 
 ## Prerender
 
-Export a `prerender` variable to prerender routes. `prerender` can be an array or set of paths to prerender, or a function that returns the paths.
+Export a `prerender` variable to prerender routes. `prerender` can be an `Array` or a `Set` of paths to prerender, or a function that returns the paths.
 
 ```ts
 // src/server/+func.ts
@@ -178,7 +172,7 @@ export const prerender: Prerender = [
 ];
 ```
 
-After the Vite build is complete, domco will import `default.fetch` from your function module and request the routes provided. The responses will be written to `dist/client/(prerender-path)` files upon build. If the path does not have an extension, `/index.html` will be added to the end of the file path to write.
+After the Vite build is complete, domco will import `default.fetch` from your function module and make a request to each prerender path provided. The responses will be written to `dist/client/(path)` files upon build. If the path does not have an extension, `/index.html` will be added to the end of the file path.
 
 For the export above, domco would request each path and generate the following files from the responses.
