+++ package.json
@@ -18,14 +18,14 @@
 	},
 	"devDependencies": {
 		"@changesets/cli": "^2.28.1",
-		"@robino/md": "^3.2.0",
+		"@robino/md": "^3.2.1",
 		"@robino/prettier": "^1.0.2",
 		"@robino/tsconfig": "^1.0.2",
-		"hono": "^4.7.4",
+		"hono": "^4.7.5",
 		"prettier": "^3.5.3",
 		"prettier-plugin-tailwindcss": "^0.6.11",
 		"turbo": "^2.4.4",
 		"typescript": "^5.8.2",
-		"vite": "^6.2.1"
+		"vite": "^6.2.3"
 	}
 }
